pyproject.toml
[project]
name = "cue-mcp"
version = "0.1.0"
description = "Computer Use Enhanced MCP Server - GUI automation with intelligent augmentation for Claude Code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "yonghwan1106"}]

dependencies = [
    "mcp>=1.26.0",
    "Pillow>=10.0.0",
    "comtypes>=1.4.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
grounding = [
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "numpy>=1.24.0",
    "scikit-image>=0.21.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cue_mcp"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
